refactor: extract tenant collection helper module
